Refactor BaGet
