+++ src/BaGet.Core.Server/Controllers/SearchController.cs
