+++ src/BaGet.Core.Server/Controllers/SymbolController.cs
