+++ src/BaGet.Core.Server/Extensions/HttpRequestExtensions.cs
