+++ src/BaGet.Core.Server/Extensions/IRouteBuilderExtensions.cs
